refactor : FeedDetailPage 리팩토링
